feat: resolver 함수 -> token 이 아닌, memberId 를 추출하여 파라미터에 주입하는 것으로 수정
