+++ src/test/java/kuit/subway/study/auth/AuthServiceMockTest.java
@@ -64,35 +64,35 @@ public class AuthServiceMockTest {
 
     }
 
-    @Nested
-    @DisplayName("로그인 후, 회원 정보 조회 Mock 테스트")
-    class GetMyMemberInfo {
-
-        @Nested
-        @DisplayName("정상 케이스")
-        class SuccessCase {
-            @Test
-            @DisplayName("로그인 시 생성된 토큰을 이용하여 내 정보 조회")
-            void getMyMemberInfoSuccess() {
-
-                // given
-                Member member = new Member(AGE, EMAIL, PASSWORD);
-                given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
-                given(jwtTokenProvider.createToken(member.getId())).willReturn("TOKEN");
-
-                // when
-                MemberResponse findMember = authService.findLoginMemberByToken("TOKEN");
-
-                // then
-                assertAll(() -> {
-                    assertThat(findMember.getEmail()).isEqualTo(EMAIL);
-                    assertThat(findMember.getEmail()).isEqualTo(EMAIL);
-                });
-
-            }
-        }
-
-    }
+//    @Nested
+//    @DisplayName("로그인 후, 회원 정보 조회 Mock 테스트")
+//    class GetMyMemberInfo {
+//
+//        @Nested
+//        @DisplayName("정상 케이스")
+//        class SuccessCase {
+//            @Test
+//            @DisplayName("로그인 시 생성된 토큰을 이용하여 내 정보 조회")
+//            void getMyMemberInfoSuccess() {
+//
+//                // given
+//                Member member = new Member(AGE, EMAIL, PASSWORD);
+//                given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
+//                given(jwtTokenProvider.createToken(member.getId())).willReturn("TOKEN");
+//
+//                // when
+//                MemberResponse findMember = authService.findLoginMemberByToken("TOKEN");
+//
+//                // then
+//                assertAll(() -> {
+//                    assertThat(findMember.getEmail()).isEqualTo(EMAIL);
+//                    assertThat(findMember.getEmail()).isEqualTo(EMAIL);
+//                });
+//
+//            }
+//        }
+//
+//    }
 
 
 }
